+++ R/package.R
@@ -3,7 +3,7 @@
 #' @import R6
 #' @importFrom reticulate import dict iterate import_from_path py_iterator py_call py_capture_output py_get_attr py_has_attr py_is_null_xptr py_to_r r_to_py tuple
 #' @importFrom graphics par plot points
-#' @importFrom tensorflow tf_version 
+#' @importFrom tensorflow tf_version tf_config install_tensorflow
 NULL
 
 # Main Keras module